# api/openapi.yaml
openapi: 3.1.0

info:
  title: gift-list
  version: 1.0.0

paths:

  # groups

  /groups:
    get:
      summary: Lists all the groups
      tags:
        - Groups
      security:
        - adminAuth: []
      responses:
        '200':
          description: The list of the groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Group'
    post:
      summary: Create a new group
      tags:
        - Groups
      security:
        - adminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroup'
      responses:
        '201':
          description: The created group.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '500':
          description: Some server error

  /groups/{name}:
    get:
      summary: Get the group by name
      tags:
        - Groups
      security:
        - adminAuth: []
      parameters:
        - in: path
          name: name
          schema:
            type: string
          required: true
          description: The group name
      responses:
        '200':
          description: The group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '404':
          description: The group was not found
    put:
      summary: Update the group by the id
      tags:
        - Groups
      security:
        - adminAuth: []
      parameters:
        - in: path
          name: name
          schema:
            type: string
          required: true
          description: The group name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGroup'
      responses:
        '204':
          description: The group was updated
        '404':
          description: The group was not found
        '500':
          description: Some error happened
    delete:
      summary: Remove the group by name
      tags:
        - Groups
      security:
        - adminAuth: []
      parameters:
        - in: path
          name: name
          schema:
            type: string
          required: true
          description: The group name
      responses:
        '204':
          description: The group was deleted
        '404':
          description: The group was not found

  # users

  /users:
    get:
      summary: Lists users in the group
      tags:
        - Users
      security:
        - userAuth: []
      responses:
        '200':
          description: The list of the user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserListItem'
    patch:
      summary: Order users
      tags:
        - Users
      security:
        - userAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/OrderUser'
      responses:
        '204':
          description: Order for group updated.
        '500':
          description: Some server error
    post:
      summary: Create a new user
      tags:
        - Users
      security:
        - userAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUser'
      responses:
        '201':
          description: The created user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '500':
          description: Some server error
  /users/{id}:
    put:
      summary: Update a user
      tags:
        - Users
      security:
        - userAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: The user id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUser'
      responses:
        '204':
          description: The user has updated
        '500':
          description: Some server error
    delete:
      summary: Remove the user by id
      tags:
        - Users
      security:
        - userAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: The user id
      responses:
        '204':
          description: The user was deleted
        '404':
          description: The user was not found

  # gifts

  /users/{userId}/gifts:
    get:
      summary: Lists the user's gifts
      tags:
        - Gifts
      security:
        - userAuth: []
      parameters:
        - in: path
          name: userId
          schema:
            type: string
          required: true
          description: The user id
      responses:
        '200':
          description: The list of the user's gifts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GiftListItem'
    patch:
      summary: Order user's gifts
      tags:
        - Gifts
      security:
        - userAuth: []
      parameters:
        - in: path
          name: userId
          schema:
            type: string
          required: true
          description: The user id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/OrderGift'
      responses:
        '204':
          description: Gifts ordered for the user.
        '403':
          description: Forbidden
        '404':
          description: Gift not found
        '500':
          description: Some server error
    post:
      summary: Create a new gift
      tags:
        - Gifts
      security:
        - userAuth: []
      parameters:
        - in: path
          name: userId
          schema:
            type: string
          required: true
          description: The user id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGift'
      responses:
        '201':
          description: The created gift.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Gift'
        '500':
          description: Some server error
  /users/{userId}/gifts/{giftId}:
    put:
      summary: Update a gift
      tags:
        - Gifts
      security:
        - userAuth: []
      parameters:
        - in: path
          name: userId
          schema:
            type: string
          required: true
          description: The user id
        - in: path
          name: giftId
          schema:
            type: string
          required: true
          description: The gift id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGift'
      responses:
        '204':
          description: The gift has updated
        '500':
          description: Some server error
    delete:
      summary: Remove the gift by id
      tags:
        - Gifts
      security:
        - userAuth: []
      parameters:
        - in: path
          name: userId
          schema:
            type: string
          required: true
          description: The user id
        - in: path
          name: giftId
          schema:
            type: string
          required: true
          description: The gift id
      responses:
        '204':
          description: The gift was deleted
        '404':
          description: The gift was not found
  /users/{userId}/gifts/{giftId}/offer:
    post:
      summary: Offer the gift
      tags:
        - Gifts
      security:
        - userAuth: []
      parameters:
        - in: path
          name: userId
          schema:
            type: string
          required: true
          description: The user id
        - in: path
          name: giftId
          schema:
            type: string
          required: true
          description: The gift id
      responses:
        '204':
          description: The gift was offered
        '403':
          description: Forbidden
        '404':
          description: The gift was not found
  /users/{userId}/gifts/{giftId}/unoffer:
    post:
      summary: Unoffer the gift
      tags:
        - Gifts
      security:
        - userAuth: []
      parameters:
        - in: path
          name: userId
          schema:
            type: string
          required: true
          description: The user id
        - in: path
          name: giftId
          schema:
            type: string
          required: true
          description: The gift id
      responses:
        '204':
          description: The gift was unoffered
        '403':
          description: Forbidden
        '404':
          description: The gift was not found

# components

components:

  # security

  securitySchemes:
    adminAuth:
      type: http
      scheme: basic
    userAuth:
      type: http
      scheme: basic

  schemas:

    # groups

    Group:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: The group name
      example:
        name: Jedi order

    CreateGroup:
      type: object
      required:
        - name
        - password
      properties:
        name:
          type: string
          description: The group name
        password:
          type: string
          description: The group password
      example:
        name: Jedi order
        password: maythe4bewithyou

    UpdateGroup:
      type: object
      required:
        - password
      properties:
        password:
          type: string
          description: The group password
      example:
        password: maythe4bewithyou

    # users

    User:
      type: object
      required:
        - id
        - name
        - picture
      properties:
        id:
          type: string
          description: The user ID
        name:
          type: string
          description: The user name
        picture:
          type: string
          description: The user picture
      example:
        id: 840eec48-5c81-4b93-9343-11d3026f244d
        name: Obiwan
        picture: light-saber

    UserListItem:
      type: object
      required:
        - id
        - name
        - picture
        - order
      properties:
        id:
          type: string
          description: The user ID
        name:
          type: string
          description: The user name
        picture:
          type: string
          description: The user picture
        order:
          type: number
      example:
        id: 840eec48-5c81-4b93-9343-11d3026f244d
        name: Obiwan
        picture: light-saber
        order: 1

    OrderUser:
      type: object
      required:
        - userId
        - order
      properties:
        userId:
          type: string
          description: The user id
        order:
          type: number
          description: The order number
      example:
        id: 840eec48-5c81-4b93-9343-11d3026f244d
        order: 1

    CreateUser:
      type: object
      required:
        - name
        - picture
      properties:
        name:
          type: string
          description: The user name
        picture:
          type: string
          description: The user picture
      example:
        name: Obiwan
        picture: light-saber

    UpdateUser:
      type: object
      required:
        - name
        - picture
      properties:
        name:
          type: string
          description: The user name
        picture:
          type: string
          description: The user picture
      example:
        name: Obiwan
        picture: light-saber

    # gifts

    Gift:
      type: object
      required:
        - id
        - name
        - userId
      properties:
        id:
          type: string
          description: The gift ID
        name:
          type: string
          description: The gift name
        link1:
          type: string
          description: The first gift link
        link2:
          type: string
          description: The second gift link
        link3:
          type: string
          description: The third gift link
        userId:
          type: string
          description: The user gift id
        offeredBy:
          type: string
          description: The id of the user giving the gift
      example:
        id: 3e548e1b-44e4-4035-a281-a5226a0196fd
        name: Light saber
        link1: https://jedi-order.io/shop/light-saber/JKF93G47
        userId: 840eec48-5c81-4b93-9343-11d3026f244d
        offeredBy: fd48dca9-fbd5-4d21-b729-da09d5a3a835

    GiftListItem:
      type: object
      required:
        - id
        - name
        - userId
        - order
      properties:
        id:
          type: string
          description: The gift ID
        name:
          type: string
          description: The gift name
        link1:
          type: string
          description: The first gift link
        link2:
          type: string
          description: The second gift link
        link3:
          type: string
          description: The third gift link
        userId:
          type: string
          description: The user gift id
        offeredBy:
          type: string
          description: The id of the user giving the gift
        order:
          type: number
      example:
        id: 3e548e1b-44e4-4035-a281-a5226a0196fd
        name: Light saber
        link1: https://jedi-order.io/shop/light-saber/JKF93G47
        userId: 840eec48-5c81-4b93-9343-11d3026f244d
        offeredBy: fd48dca9-fbd5-4d21-b729-da09d5a3a835
        order: 1

    OrderGift:
      type: object
      required:
        - giftId
        - order
      properties:
        giftId:
          type: string
          description: The gift id
        order:
          type: number
          description: The order number
      example:
        id: 3e548e1b-44e4-4035-a281-a5226a0196fd
        order: 1

    CreateGift:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: The user name
        link1:
          type: string
          description: The first gift link
        link2:
          type: string
          description: The second gift link
        link3:
          type: string
          description: The third gift link
      example:
        name: Light saber
        link1: https://jedi-order.io/shop/light-saber/JKF93G47

    UpdateGift:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: The user name
        link1:
          type: string
          description: The first gift link
        link2:
          type: string
          description: The second gift link
        link3:
          type: string
          description: The third gift link
        userId:
          type: string
          description: The user gift id
      example:
        name: Light saber
        link1: https://jedi-order.io/shop/light-saber/JKF93G47

tags:
  - name: Groups
    description: The groups
  - name: Users
    description: The users
  - name: Gifts
    description: The gifts